Move tests to unit folder and update imports
